+++ CHANGELOG.md
@@ -339,3 +339,8 @@
   - Добавлен BackgroundCycle
   - Исправлены запросы к микросервису
   - В BackgroundCycle добавлено обновление проверки работы сервера
+
+5-2-2026:
+  - Полностью изменен датасет и сбор данных для него
+  - В загрузке данных у модели теперь фильтруются NaN значения
+  - collect_all теперь метод CCXTClient, collect_from_slice теперь CollectedData::from_slice(), flat_all теперь FlattenedData::from_collected()